Drawer Overhaul
Drawer now draws based on circles built by the Canvas class. Each circle
is known as a Draw Point.

DrawEditor now takes a boolean so that it can be either the drawer or
just set up the necessary classes for a guesser.

Encoding scheme for DrawPoints in progress. Same goes for how updates
should be packaged by the drawer.